apache camel propagation tests - cleanup #1
