+++ instrumentation/apache-camel-2.20/javaagent/src/test/groovy/io/opentelemetry/javaagent/instrumentation/apachecamel/aws/SnsConfig.groovy
@@ -3,16 +3,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.apachecamel
+package io.opentelemetry.javaagent.instrumentation.apachecamel.aws
 
-import com.amazonaws.services.s3.AmazonS3Client
-import com.amazonaws.services.sns.AmazonSNSAsyncClient
-import com.amazonaws.services.sns.AmazonSNSClient
-import com.amazonaws.services.sqs.AmazonSQSAsyncClient
-import com.amazonaws.services.sqs.AmazonSQSClient
+
 import org.apache.camel.LoggingLevel
 import org.apache.camel.builder.RouteBuilder
-import org.apache.camel.component.aws.s3.S3Constants
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.boot.SpringBootConfiguration
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration
@@ -20,10 +15,10 @@ import org.springframework.context.annotation.Bean
 
 @SpringBootConfiguration
 @EnableAutoConfiguration
-class S3Config {
+class SnsConfig {
 
   @Bean
-  RouteBuilder sqsDirectlyFromS3ConsumerRoute(@Value("\${queueName}") String queueName) {
+  RouteBuilder sqsCamelOnlyConsumerRoute(@Value("\${queueName}") String queueName) {
     return new RouteBuilder() {
 
       @Override
@@ -36,7 +31,7 @@ class S3Config {
   }
 
   @Bean
-  RouteBuilder s3ToSqsProducerRoute(@Value("\${bucketName}") String bucketName) {
+  RouteBuilder snsProducerRoute(@Value("\${topicName}") String topicName) {
     return new RouteBuilder() {
 
       @Override
@@ -44,27 +39,8 @@ class S3Config {
         from("direct:input")
           .log(LoggingLevel.INFO, "test", "SENDING body: \${body}")
           .log(LoggingLevel.INFO, "test", "SENDING headers: \${headers}")
-          .convertBodyTo(byte[].class)
-          .setHeader(S3Constants.KEY,simple("test-data"))
-          .to("aws-s3://${bucketName}?amazonS3Client=#s3Client")
+          .to("aws-sns://${topicName}?amazonSNSClient=#snsClient")
       }
     }
-  }
-
-  @Bean
-  AmazonSQSClient sqsClient() {
-    return AmazonSQSAsyncClient.asyncBuilder()
-      .build()
-  }
-
-  @Bean
-  AmazonSNSClient snsClient() {
-    return AmazonSNSAsyncClient.asyncBuilder()
-      .build()
-  }
-
-  @Bean
-  AmazonS3Client s3Client() {
-    return AmazonS3Client.builder().build()
   }
 }
